move only test in test_dace_lazy_stencil to test_dace_parsing, fixes CI issue
